More warnings are fixes
Now, also all tigl::unique_ptr usages should be removed
and replaced by std::unique_ptr.

+++ src/math/CTiglProjectPointOnCurveAtAngle.cpp
@@ -187,7 +187,7 @@ int tigl::CTiglProjectPointOnCurveAtAngle::NbPoints() const
 {
     Compute();
 
-    return resultParameter.size();
+    return static_cast<int>(resultParameter.size());
 }
 
 void CTiglProjectPointOnCurveAtAngle::CTiglProjectPointOnCurveAtAngle::Compute() const
